refactor: more structured backend, added docker compose and .env
