+++ .example.env
@@ -1,0 +1,5 @@
+MONGO_INITDB_ROOT_USERNAME=
+MONGO_INITDB_ROOT_PASSWORD=
+MONGO_INITDB_DATABASE=
+MONGO_URI=
+BACKEND_PORT=
