+++ .gitignore
@@ -52,3 +52,6 @@ Thumbs.db
 .svn
 .svn/
 CVS
+
+# enviroment files
+.env
