+++ backend/src/models/user.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct User {
+    #[serde(rename = "_id")]
+    pub id: u32,
+    pub username: String,
+    pub password: String,
+}
